refactor: extract CM-related snippet functions into new file
